+++ bot/actions/action_telefone.py
@@ -1,4 +1,4 @@
-# Este arquivo contém custom actions que utilizão código python
+# Este arquivo contém uma custom action que utiliza código python
 # para executar ações no diálogo.
 #
 # Veja o guia na documentação do RASA em:
@@ -9,24 +9,6 @@ from typing import Any, Text, Dict, List
 from rasa_sdk import Action, Tracker
 from rasa_sdk.executor import CollectingDispatcher
 from rasa_sdk.events import SlotSet
-
-
-class ActionTeste(Action):
-    def name(self) -> Text:
-        return "action_teste"
-
-    def run(
-        self,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: Dict[Text, Any],
-    ) -> List[Dict[Text, Any]]:
-        try:
-            dispatcher.utter_message("Mensagem enviada por uma custom action.")
-        except ValueError:
-            dispatcher.utter_message(ValueError)
-        return []
-
 
 class ActionTelefone(Action):
     def name(self) -> Text:
