+++ bot/actions/action_teste.py
@@ -1,0 +1,26 @@
+# Este arquivo contém uma custom action que utiliza código python
+# para executar ações no diálogo.
+#
+# Veja o guia na documentação do RASA em:
+# https://rasa.com/docs/rasa/core/actions/#custom-actions/
+
+from typing import Any, Text, Dict, List
+
+from rasa_sdk import Action, Tracker
+from rasa_sdk.executor import CollectingDispatcher
+
+class ActionTeste(Action):
+    def name(self) -> Text:
+        return "action_teste"
+
+    def run(
+        self,
+        dispatcher: CollectingDispatcher,
+        tracker: Tracker,
+        domain: Dict[Text, Any],
+    ) -> List[Dict[Text, Any]]:
+        try:
+            dispatcher.utter_message("Mensagem enviada por uma custom action.")
+        except ValueError:
+            dispatcher.utter_message(ValueError)
+        return []
